refactor navigation
